MoveCamera must switch to perspective camera mode
TransitionCamera::MoveCamera is used to transition to any arbitrary perspective viewpoint. BUT if the camera is currently in orthographic mode, it must first switch to perspective camera mode before transitioning to the target viewpoint.

This includes constraining 'yaw' to be within 0 to 360 degrees and finding the shortest direction to transition in.  This will avoid any 'unwinding' behavior.